I wrote a small test program to measure the differences in runtime. Reading in and reverse-complementing 1 million 150 bp reads takes 4.5 µs per read with the previous code and 2.2 µs per read with this version using a lookup table. Just parsing the FASTQ takes 1.2 µs, so reverse complementing time goes from 3.3 µs to 1.0 µs per read.

I tried to measure what the effect is on mapping time, but it makes nearly no difference (less than 1%). For the fruit fly dataset, mapping time per read went from something like 194.7±0.5µs before to 194.0±0.5 µs after the change.

It appears to me that reverse complementing is not responsible for any significant part of the mapping runtime. We currently have a caching mechanism for reverse complements, but I’m tempted to remove that and just *always* compute each read’s reverse complement because the caching itself complicates the code and also makes it slower.

Closes #174