@ksahlin 2023-01-25
Hmm okay. Just to verify: 

We have a default maximum of M locations to try for each read (default 20). In worst case for a read, all M locations are tried and all of them are reverse complements. This would mean with your new code it would be Mµs for a read instead of the old (3.3+ (M-1)X)µs per read (X is the hash table lookup time for the stored rev comp read). 

Since the large majority of read will not hit this worst case I guess it is fine. There is definitely value in making the code simple.

You have my approval to do as you wish :)