@marcelm 2023-01-25
No, we would just always keep the read and its reverse complement in memory. Then when the match is to the reverse strand, the reverse version is used and otherwise the normal one. Or do I misunderstand?