@ksahlin 2023-01-25
Sorry, I misunderstood. Long day.. Attempt 2:

Ok I see, it is in some way the opposite of what I wrote. Each read will have both directions computed and stored. We know that at least half of the reads need to be reversed regardless in PE mode. So, compared computing RC only on a need case basis, it will in 'worst case'  be an extra X/2µs for computing RC for X reads if there are X read-pairs (this happens if all reads have a unique location).

Then, isen't what I wrote above both simpler (or as simple) and faster (assuming that the large majority of reads will have unique locations)? That is, to obtain the FW or RC string whenever needed with potential of repetitive calls?

Anyways, since this does not seem to affect runtime in anyway, I am happy with what you proposed, to avoid starting another investigation of these two approaches. You have the executive decision.