@marcelm 2023-01-26
> Then, isen't what I wrote above both simpler (or as simple) and faster (assuming that the large majority of reads will have unique locations)? That is, to obtain the FW or RC string whenever needed with potential of repetitive calls?

We already have a [`Read` class](https://github.com/ksahlin/strobealign/blob/70b3f51ace8a478e14a4aad102e7ef24a098dd8a/src/revcomp.hpp#L24) that is used consistently for getting either the forward or reverse sequence of a read. It has a `.rc()` method that computes the reverse complement on first access. Version 0.7.1 did the same thing, but the logic was not encapsulated in a separate class. I refactored this in 4b99f3945e7673843d5319cdaa36d3545bfa8356. The benefit with this is that we don’t need to compute the reverse complement if it turns out that we never need it, but the downside is that there is the overhead of checking whether we have reverse-complemented or not.

I’ll push a commit shortly that just always computes the reverse complement when the `Read` is constructed. This actually saves another half microsecond per read.

> Anyways, since this does not seem to affect runtime in anyway, I am happy with what you proposed, to avoid starting another investigation of these two approaches. You have the executive decision.

Ok, I’ll push the mentioned commit and then merge this!